pyproject.toml
[project]
name = "pydanticai-explore"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.193.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "pydantic-ai>=1.70.0",
    "streamlit>=1.55.0",
]
